fix: 修复 CI 编译错误（SSE 重连延迟类型不匹配 + DiffCard 高亮 @Composable 标注）
- ReasonixSseClient: retryWhen 的 attempt 为 Long，coerceAtMost(5L).toInt() 后再用于 shl 位移，指数退避语义不变
- DiffCard: 新增 DiffHighlightColors 颜色快照，在 @Composable 上下文读取 5 个颜色后传入非 composable 的 highlightDiff/appendNodes/fallbackHighlight，避免在非 composable 函数中访问 @Composable getter，并保留 remember 缓存

+++ app/src/main/java/com/reasonix/agents/data/api/ReasonixSseClient.kt
@@ -131,7 +131,7 @@ class ReasonixSseClient(
             false
         } else {
             // attempt: 0→1s, 1→2s, 2→4s, 3→8s, 4→16s, ≥5→30s（封顶）
-            val delayMs = (1000L shl attempt.coerceAtMost(5)).coerceAtMost(30_000L)
+            val delayMs = (1000L shl attempt.coerceAtMost(5L).toInt()).coerceAtMost(30_000L)
             delay(delayMs)
             true
         }
